Clean pedo trash code, fix service stop when phone was locked
