+++ app/src/main/AndroidManifest.xml
@@ -8,6 +8,8 @@
     <uses-permission android:name="android.permission.MODIFY_AUDIO_SETTINGS" ></uses-permission>
     <uses-permission android:name="android.permission.VIBRATE"></uses-permission>
     <uses-permission android:name="android.permission.ACTIVITY_RECOGNITION"></uses-permission>
+    <uses-permission android:name="android.permission.WAKE_LOCK" />
+
 
     <application
         android:allowBackup="true"
@@ -19,12 +21,12 @@
         <activity android:name=".ranking.RankingActivity"></activity>
 
         <service
-            android:name=".Test.StepService3"
+            android:name=".Pedometer.StepService3"
             android:enabled="true"
             android:exported="true" />
 
         <service
-            android:name=".Test.CheckAppRunService"
+            android:name=".Pedometer.CheckAppRunService"
             android:enabled="true"
             android:exported="true" />
 
@@ -33,7 +35,7 @@
             android:enabled="true"
             android:exported="true" />
 
-        <receiver android:name=".Test.NotiReceiver" />
+        <receiver android:name=".Pedometer.MeetGoalReceiver" />
 
         <activity
             android:name=".stepcounter.ProgressActivity"
