+++ app/src/main/java/neu/madcourse/walkwithme/Pedometer/Constants.java
@@ -1,8 +1,6 @@
-package neu.madcourse.walkwithme.Test;
+package neu.madcourse.walkwithme.Pedometer;
 
 public class Constants {
     public static final String START_FOREGROUND =  "start_foreground";
     public static final String STOP_FOREGROUND = "stop_foreground";
-    public static final String RESET_COUNT = "reset_count";
-    public static final String STOP_SAVE_COUNT = "stop_save_count";
 }
