--- app/src/main/java/neu/madcourse/walkwithme/Test/NofiticationConstants.java
@@ -1,28 +1,0 @@
-package neu.madcourse.walkwithme.Test;
-
-
-import neu.madcourse.walkwithme.R;
-
-public class NofiticationConstants{
-    public static final NotificationMessage L2 = new NotificationMessage("Congratulations \uD83C\uDF8A \uD83C\uDF8A","You've leveled up to level 2!", 1, R.drawable.fireworks);
-    public static final NotificationMessage L3 = new NotificationMessage("Hooray \uD83C\uDF89 \uD83C\uDF89", "That's level 3!", 2, R.drawable.fireworks);
-    public static final NotificationMessage L4 = new NotificationMessage("Woo Hoo \uD83D\uDC4F \uD83D\uDC4F \uD83D\uDC4F", "You made it to level 4!", 3, R.drawable.fireworks);
-    public static final NotificationMessage L5 = new NotificationMessage("❤Amazing❤", "You've reached level 5!", 4, R.drawable.fireworks);
-
-    public static final NotificationMessage b1elowGoal = new NotificationMessage("\uD83D\uDD08 Almost there", "Only few steps away from your daily goal! \uD83D\uDEB6\u200D♂️ \uD83D\uDEB6\u200D♂️", 5, R.drawable.happy);
-
-    public static final NotificationMessage b1elowGoal2 = new NotificationMessage("\uD83D\uDE30 Looks like you forgot....", "It't not too late to meet your daily goal!️", 5, R.drawable.sleep);
-
-
-
-
-
-    public static final String[] LEVEL_3 =  {"Hooray! That's level 3!", "1"};
-    public static final String[] LEVEL_4 =  {"Woo Hoo! You made it to level 4!", "1"};
-    public static final String[] LEVEL_5 =  {"Amazing!! You've reached level 5!", "1"};
-
-// NotifiContst.L2.msg
-    public static final String REACH_DAILY_GOAL = "Excellent! You have met your daily goal!";
-    public static final String THREE_DAYS_INACTIVE = "Uh oh, We miss you! Why don't you check what is going in your app?";
-
-}
