--- app/src/main/java/neu/madcourse/walkwithme/stepcounter/StepListener.java
@@ -1,5 +1,0 @@
-package neu.madcourse.walkwithme.stepcounter;
-
-interface StepListener {
-    public void step(long timeNs);
-}
