+++ app/src/main/java/neu/madcourse/walkwithme/userlog/LoginActivity.java
@@ -12,7 +12,6 @@ import android.widget.EditText;
 import android.widget.Toast;
 
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -20,9 +19,7 @@ import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 
 import neu.madcourse.walkwithme.MainActivity;
-import neu.madcourse.walkwithme.Test.Constants;
-import neu.madcourse.walkwithme.Test.StepService3;
-import neu.madcourse.walkwithme.profile.ProfileActivity;
+import neu.madcourse.walkwithme.Pedometer.StepService3;
 import neu.madcourse.walkwithme.R;
 import neu.madcourse.walkwithme.ranking.DRankingData;
