--- app/src/main/res/layout/fragment_steps.xml
@@ -1,115 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
->
-
-    <ScrollView
-        android:layout_width="match_parent"
-        android:layout_height="578dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginStart="2dp"
-        android:layout_marginTop="3dp"
-        android:layout_marginEnd="0dp">
-
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:paddingHorizontal="16dp">
-
-
-            <TextView
-                android:id="@+id/dayRecordText"
-                android:layout_width="246dp"
-                android:layout_height="match_parent"
-                android:layout_alignParentStart="true"
-                android:layout_alignParentTop="true"
-                android:layout_marginStart="21dp"
-                android:layout_marginTop="11dp"
-                android:layout_weight="1"
-                android:autoSizeMaxTextSize="30dp"
-                android:autoSizeMinTextSize="2sp"
-                android:autoSizeTextType="uniform"
-                android:gravity="center_vertical"
-                android:padding="5dp"
-                android:textSize="22sp"
-                tools:text="Today's Goal : " />
-
-            <Button
-                android:id="@+id/setgoal"
-                android:layout_width="100dp"
-                android:layout_height="match_parent"
-                android:layout_alignParentTop="true"
-                android:layout_alignParentEnd="true"
-                android:layout_marginTop="5dp"
-                android:layout_marginEnd="8dp"
-                android:autoSizeMaxTextSize="30dp"
-                android:autoSizeMinTextSize="2sp"
-                android:autoSizeTextType="uniform"
-                android:maxLines="1"
-                android:padding="5dp"
-                android:text="Set Goal"
-                android:textAllCaps="false"
-                android:textSize="20dp"
-                android:textStyle="bold" />
-
-            <TextView
-                android:id="@+id/stepText"
-                android:layout_width="345dp"
-                android:layout_height="185dp"
-                android:layout_below="@+id/stepLabel"
-                android:layout_alignParentStart="true"
-                android:layout_marginStart="47dp"
-                android:layout_marginTop="72dp"
-                android:gravity="center"
-                android:textSize="60sp"
-                android:textStyle="bold"
-                tools:text="0" />
-
-            <TextView
-                android:id="@+id/accuracy_alert"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_below="@id/stepText"
-                android:layout_alignParentStart="true"
-                android:layout_marginStart="8dp"
-                android:layout_marginTop="8dp"
-                android:padding="5dp"
-                android:text="Alert will be dsplayed here"
-                android:textStyle="bold" />
-
-
-        </RelativeLayout>
-    </ScrollView>
-
-    <Button
-        android:id="@+id/resetButton"
-        android:layout_width="107dp"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginStart="16dp"
-        android:layout_marginLeft="16dp"
-        android:layout_marginBottom="181dp"
-        android:text="Reset"
-        android:textAllCaps="false"
-        android:textStyle="bold" />
-
-    <Button
-        android:id="@+id/startButton"
-        android:layout_width="98dp"
-        android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginEnd="29dp"
-        android:layout_marginRight="16dp"
-        android:layout_marginBottom="183dp"
-        android:text="Start"
-        android:textAllCaps="false"
-        android:textStyle="bold" />
-
-</RelativeLayout>
